Allow admins to disable branding and assurance messages.
This will allow the administrator of a server to disable the prefix
"Prism //" if they want to, as the player and moderator know they're
using Prism and the brightly coloured message can be distracting when
focusing on a task such as rolling back. It will also allow the admin
to disable assurance messages such as "Like a boss.", "It's like it
never happened.", and "It's like it was always there." for the same
reasons as above - they can be distracting. This will enable server
admins to apply a more minimalisitic feel to Prism with messages such
as "67 reverted. 54 ignored." instead of "Prism // 67 reverted. 54
ignored. It's like it never happened."