Recording task now reschedules itself. This should allow us to be much more versatile with database connections in that:
- Insert batches will never overlap because the next task isn't scheduled until this one is complete
- If there are too many connections errors or dead connections we can better handle waiting. The scheduler will keep increase the time we wait so the system doesn't go crazy with connection attempts and console spam.
- This *should* solve the error in which extra data inserts are not properly lining up in a small set of queries. More testing is needed to confirm.

The system still needs a way to handle unsaved data if the db never comes back online, but in tests with the db shut down and restored, everything works well.

Also made some minor cleanup related to this... the config no longer defines how long to wait. The system properly converts seconds to wait to ticks, and the extra data is included in the original batch connection transaction